requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
python-dotenv 
python-jose[cryptography]
passlib[argon2]
python-multipart
pydantic
requests
huggingface_hub